minor styling changes in product
