+++ client/src/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import Navbar from './components/layout/Navbar';
 import Landing from './components/layout/Landing';
 import Routes from './components/routing/Routes';
-import { ProductProvider } from "../src/context";
+import { ProductProvider } from "../src/utils/context";
 
 // Redux
 import { Provider } from 'react-redux';
