+++ client/src/components/product/Details.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { ProductConsumer } from "../../context";
+import { ProductConsumer } from "../../utils/context";
 import { Link } from "react-router-dom";
 import { ButtonContainer } from "./Button";
